fix and reorganize SparkFileDataObjectTest
